Changed endpoints routes
