+++ EleksInternshipProj.Server/EleksInternshipProj.WebApi/Controllers/TaskController.cs
@@ -79,7 +79,6 @@ namespace EleksInternshipProj.WebApi.Controllers
         }
 
         [HttpPost]
-        [Route("create")]
         public async Task<IActionResult> Create([FromBody] TaskDto dto)
         {
             try
@@ -131,7 +130,7 @@ namespace EleksInternshipProj.WebApi.Controllers
         }
 
         [HttpDelete]
-        [Route("delete/{id:long}")]
+        [Route("{id:long}")]
         public async Task<IActionResult> Delete(long id)
         {
             try
